modbus_server/requirements.txt
pymodbus
python-dotenv
requests>=2.0